split functions into components
