+++ src/app/shared/redux/type.ts
@@ -2,4 +2,4 @@ export const TODO_ADD = 'TODO_ADD';
 export const TODO_DELETE = 'TODO_DELETE';
 export const TODO_UPDATE = 'TODO_UPDATE';
 export const TODO_CLEAR = 'TODO_CLEAR';
-export const TODO_COMPLETE = 'TODO_COMPLETE';
+export const TODO_TOGGLE_STATUS = 'TODO_TOGGLE_STATUS';
