+++ src/app/stylesheets/pages/_home.scss
@@ -1,5 +1,5 @@
 .App {
-  width: 100vw;
+  width: 100%;
   height: 100vh;
   background: linear-gradient(180deg, rgba(0, 0, 0, 0.8), rgba(0, 0, 0, 0.5)),
     url(https://images.unsplash.com/photo-1523374228107-6e44bd2b524e?auto=format&fit=crop&ixid=eyJhcHBfaWQiOjEyMDd9&ixlib=rb-1.2.1&q=80&w=1350)
@@ -15,7 +15,7 @@
     color: map-get($map: $color, $key: light);
   }
 
-  .todo-list {
+  .todo-wrapper {
     padding: 2rem 2rem 1rem;
     background-color: hsla(0, 0%, 96.1%, 0.9);
     box-shadow: 0 1px 1px rgba(0, 0, 0, 0.2), 0 8px 0 -3px #f5f5f5,
@@ -36,6 +36,10 @@
     border-bottom: 1px solid rgb(204, 204, 204);
   }
 
+  .todo-list {
+    max-height: 300px;
+    overflow-y: auto;
+  }
   .todo-item {
     height: 62px;
     padding: 20px 12px;
@@ -51,6 +55,9 @@
     .todo-check {
       height: 20px;
       width: 20px;
+      &:hover {
+        cursor: pointer;
+      }
     }
 
     .todo-name,
